Unify movement action into central source of truth
